+++ packages/widgetbook/lib/src/routing/routing.dart
@@ -2,4 +2,3 @@ export 'app_route_config.dart';
 export 'app_route_parser.dart';
 export 'app_router.dart';
 export 'app_router_delegate.dart';
-export 'widgetbook_shell.dart';
