+++ packages/widgetbook/test/src/layout/responsive_layout_test.dart
@@ -1,5 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:widgetbook/src/layout/desktop_layout.dart';
+import 'package:widgetbook/src/layout/layout.dart';
+import 'package:widgetbook/src/layout/mobile_layout.dart';
 import 'package:widgetbook/src/navigation/navigation.dart';
 import 'package:widgetbook/src/routing/routing.dart';
 import 'package:widgetbook/src/settings/settings.dart';
@@ -201,73 +204,6 @@ void main() {
             find.byType(SettingsPanel, skipOffstage: false),
             findsOneWidget,
           );
-
-          addTearDown(tester.view.resetPhysicalSize);
-        },
-      );
-
-      testWidgets(
-        'given a small screen, '
-        'when a four-finger tap is performed, '
-        'then the bottom bar visibility is toggled',
-        (tester) async {
-          tester.view.physicalSize = const Size(400, 800);
-          tester.view.devicePixelRatio = 1.0;
-
-          final state = WidgetbookState(
-            appBuilder: materialAppBuilder,
-            root: WidgetbookRoot(
-              children: [
-                WidgetbookFolder(
-                  name: 'Widgets',
-                  children: [],
-                ),
-              ],
-            ),
-          );
-
-          final router = AppRouter(
-            state: state,
-          );
-
-          await tester.pumpWidget(
-            WidgetbookScope(
-              state: state,
-              child: MaterialApp.router(
-                routerConfig: router,
-              ),
-            ),
-          );
-
-          final mobileShellState =
-              tester.state(find.byType(MobileWidgetbookShell))
-                  as MobileWidgetbookShellState;
-
-          expect(mobileShellState.isBottomBarVisible, isTrue);
-
-          final pointerDownEvent = const PointerDownEvent(
-            pointer: 1,
-            device: 1,
-          );
-
-          final pointerUpEvent = const PointerUpEvent(
-            pointer: 1,
-            device: 1,
-          );
-
-          mobileShellState.handlePointerDown(pointerDownEvent);
-          mobileShellState.handlePointerDown(pointerDownEvent);
-          mobileShellState.handlePointerDown(pointerDownEvent);
-          mobileShellState.handlePointerDown(pointerDownEvent);
-
-          expect(mobileShellState.isBottomBarVisible, isFalse);
-
-          mobileShellState.handlePointerUp(pointerUpEvent);
-          mobileShellState.handlePointerUp(pointerUpEvent);
-          mobileShellState.handlePointerUp(pointerUpEvent);
-          mobileShellState.handlePointerUp(pointerUpEvent);
-
-          expect(mobileShellState.isBottomBarVisible, isTrue);
 
           addTearDown(tester.view.resetPhysicalSize);
         },
